Add Assistant Home button support in Launcher TAPL
When using apps that switch to car mode (e.g. Maps) while in 3 button
navigation mode, the default navbar is not visible and instead the
overlay from Assistant is being rendered that also has the Home button.

For tests that involve apps switching to car mode, when we want to go
home we want to be able to press that button.

Bug: 282758103
Test: n/a

Change-Id: I88c420f321dcfd39164e20e97d39f90bf1e5d8d8