Update scale logic for desktop tasks in recents animation
When swiping up to overview, we need to calculate a target position for
the desktop tasks. They need to be scaled down and positioned within the
desktop tile.
After recents animation finishes, the tasks will be replaced by
screenshots. To make the replace less flickery, the position of the
tasks should match the position of the screenshots.

TODO: there is a flicker when the task is replaced with recents entry.
      Task is hidden and then, half a second later, the recents tile is
      shown. Enabling live tiles should help with this.

Bug: 298252273
Test: swipe up to recents from desktop
Change-Id: I0c7fba40dc2511c17d0011d0516456de07cfa004