Fix indexoutofbounds folder preview
The indexoutofbounds issue came from the fact that folder was being animated open when there are no items in preview. It might be a race condition where we try to open the folder at a time where it's actually being emptied. This check should prevent folders from running animateOpen() if there's <= 1 items, avoiding the outofbounds error

Fix: 298740874
Flag: no flag
Test: no test
Change-Id: I31ed53fbdf22a7f19bc9ec414242312afe7c6cd2