Starting to tighten leak detector checks
This will help to detect leaks earlier, not when we have 3 activities in some cases, but when we have 2.

Adding a parameter that enables more strict detection, and using it in some places.

I'll try to convert as many places as possible to the more strict check after landing this CL.

Bug: 187761685
Test: presubmit
Flag: N/A
Change-Id: Idd5a0386005e46dad8aab222bd77c0240f23f43a