Add custom extra width to trigger LPNH
This is defined in dp from the edge of the nav handle. So if you
set it to 48, you can invoke 48dp from the left or right of the
handle (which is roughly the size of an average fingertip).

You can also set it to a negative value to make the invocation
region smaller. This applies to both navbar and stashed taskbar,
and the minimum touch target is always 48dp.

For reference, the navbar width is 108dp and taskbar is 220dp.

Defaults to 0, i.e. you can only invoke within the visible navbar.

Bug: 325118077
Flag: LEGACY CUSTOM_LPNH_THRESHOLDS DISABLED
Test: Manual
Change-Id: I904a484a99ac4af05de13573fac610b84fd7f0f1