Further tightening leak detector checks
Some checks see 2 activities for some reason: Launcher and Overview.
Both active.
I'll have to decide what to do with these cases, but for now, I simply don't pass 'true' from places where this can happen.

When we have 2 activities:
Either Overview is present (and I guess we have to close it before the check) or the screenshot shows that there is no Overview, but internally, apparently, Overview activity is active and therefore, technically, is not leaked.
For the latter case, I may need to do some kind of investigation.
But for now, I'm tightening the check where there is no Overview.

Bug: 187761685
Test: presubmit
Flag: N/A
Change-Id: I1f1619093e0948b8112f29fd7e94facbf7093983