Developer options should reflect the correct DeviceConfig value for ReleaseFlag.
Before this change : When a ReleaseFlag is turned on from server side or
set in DeviceConfig, developer options shows the flag is off.

After this change : When a ReleaseFlag is turned on from server side or
set in DeviceConfig, developer options shows the flag is on.

This change doesn't reflect the DeviceConfig value changes done using
`adb shell device_config put` unless launcher is restarted.

Bug: 297202824
Test: Manual
Flag: NA

Change-Id: I5d83207cf454c60278fd02e59199207e855d51a8