Account for freeform tasks being visible in overview
When live tiles are enabled, desktop tasks remain visible while in
overview. This is due to tasks remaining in running state.
Update DesktopVsibilityController to handle this case. If in freeform,
it should return that freeform tasks are not visible to any callers.
Also move launcher state handling to inside the controller so we don't
have this handling in other places.
Introduce separate handling of background and overview state. In
background state we want to keep the freeform task visibility state
matching the actualy visibility. And show the launcher views and resume
the activity.
In overview state we want to override the freeform visibility and show
launcher views.

Bug: 297590571
Test: open an app on desktop
      - swipe up for taskbar multiple times, observe that we remain on
        desktop
      - swipe up for taskbar and launch an app, observe it launches on
        desktop
      - swipe up to recents, observe that recents UI is shown and task
        remains running
      - swipe up to recents, launch desktop again from recents
      - swipe up to recents, launch another fullscreen app from recents
      - swipe up from a fullscreen, launch desktop from recents
Change-Id: Id5cef7111da6929c8435ef9d221db1abd8361b15