Fix scrim attachment and detachment issues.
Attaching: The surface control from Launcher's ViewRootImpl is
occiasionally invalid when back gesture starts, and parenting scrim to
it would fail. We work around this by retrying later on back progressed.

Dettaching: The scrim is not detached in all code paths. We now always
check and detach in finishAnimation() if needed.

This also fixes a possible flicker at the beginning of the transition.
We wait until Launcher is started (and draws at least one frame) to animate the top window.

Test: m -j
Test: Repeatedly trigger back to home animation. Make sure the scrim
attaches and detaches always.

Bug:271606843
Bug:297808832
Bug:217988970
Fixes:271606843
Fixes:297808832
Fixes:217988970

Change-Id: I8b9b023f1fd54bcedd0be15196165e19e8400af7