Do DbMigration if coming from a DB w/ version < 31
This was the issue that is causing the launcher to crash when changing a
grid layout. The migration is not happening, causing there to be 21
columns in the database. When the grid size changes, and the db from
grid size X is migrated to a db of grid size Y, the column count isn't
matching.

Bug: 304687723
Test: Verified behavior through unit test.
Change-Id: I2ee1f2fe09db432715b22c41581a77f6b0ac8320