Call SetSummary for seekbarPref instead of calling onPreferenceChange().
This way the preference will not be created with the default value
initially and preference is created only when seekbar is modified
manually in developer options. Preference can be set using device_config and the default
value can be read from device config flags.

Bug: 301680992
Flag: Legacy CUSTOM_LPNH_THRESHOLDS DISABLED
Test: Manual
Change-Id: Ie48384e4799cfe0f58a469bcb7aafb971beaea19