Translucent apps are kept in Recents while in the foreground
Translucent apps that are exluded from Recents are not kept in Recents
while they are the foreground. This is to prevent jank while going to
Recents and thus morphing to another app, and to prevent confusion as
going back to an app might end you back into the excluded app experience
in case of an app like Omnient that is "on-top" of another application.

Fix: 305650114
Test: Manual
Flag: LEGACY INVOKE_OMNI_LPH DISABLED
Change-Id: I24d583434ce62c68436e8928cf3a47a5ce73ecaf