Refactor/clean up Overview Actions buttons
Since the number of Overview actions is growing, this patch factors out the button hiding logic so it's not all handled in OverviewActionsView.

Also, since splitscreen can now be carried out on Taskbar targets, the split button is no longer ever disabled (only hidden), so removed unused code for that.

Bug: 274835596
Test: Manual
Flag: NA
Change-Id: Icacdee6d0e6071ccb400dfc25a84572e7e9ce772