Making keyguard dismissal a permanent solution
Before, it was considered to be a workaround while System UI has
keyguard bugs. But these bugs will happen in future, and they are
unlikely to be prevented by presubmit tests since they are usually flakes.

The solution that this CL makes permanent is: in presubmit attempt
to unlock SysUI if it's locked. In postsubmit, just make sure it's
unlocked. This way, in postsubmit, we'll know whether tere are active
keyguard bugs.

Bug: 187761685
Test: presubmit
Flag: N/A
Change-Id: Ia9602bc6e9cc8400e939051c45aba92bc197d428