Allowing ACTION_CANCEL events to come from device 0
Tests assert that input events come to Launcher only from tests, which
sets deviceId to 0. This way, we spot when a cable touches a lab device
and breaks the tests, in which case deviceId will be != -1.

But CANCEL event, which is programmatic, now can legitimately come from
device 0.

Bug: 308489875
Test: presubmit
Flag: N/A
Change-Id: I84ee9ed8e008043f83e82d57b17b64a10677ee8d