Increasing wait time for Launcher restart after enabling test provider
Of the the logcats indicates that Launcher gets killed eventually, but
it takes 17 sec, while the test waits for the Launcher restart to
compete in 10 sec.

Perhaps restart take longer than usual because the system is still busy
killing setup wizard.

This is the second attempt to land this CL; the first one was reverted
because the revert author believed this breaks their tests.

This time I'm icreasing the wait from 10 to 30 (not 60) secs.

Bug: 308489613
Test: presubmit
Flag: N/A
Change-Id: I18cff219cf6ec2732ed181c970cb5315bf568cab