Pass through both the widget's pendingIntent and remoteView intent
* Both may be necessary to correctly launch a widget
* We pass it through as the second app's Bundle instead
of adding a new parameter.

Bug: 311263727
Test: Tested w/ 2 fresh launches, app + widget.
Also with existing task + widget.
Was not able to accurately test multiple tasks launched by the
widget since there are some issues with the allowlist check.
And the working case for multiple instances of Chrome didn't
work because none of the chrome widgets rely on the secondary
intent to launch any of the app's activities.
Flag: None, technically this code isn't reachable since widgets
are unsupported, but didn't seem worthwhile to add a separate flag
in the short duration we will begin supporting them again.
To run replace the early return in QuickstepInteractionHandler with
the call:
mLauncher.getSplitToWorkspaceController()
	.handleSecondWidgetSelectionForSplit(view, pendingIntent, options.first)

Change-Id: I1ebcb385b140d42a6a84368f589aba06c4bed2bd