Enabling ENABLE_NO_LONG_PRESS_DRAG
This is a simple flag, I don't think it requires the new
flagging system, I mainly added the flag for ease of debugging
if it breaks test on postsubmit.

This flag was already tested using enabled but it was disabled
before trunk stable, I think we can set this to true without risk
and then remove the flag.

Flag: LEGACY ENABLE_NO_LONG_PRESS_DRAG ENABLED
Fix: 299748096
Test: atest TaplDragTest
Change-Id: If99027d8cec7d75eee11012615da0e0d76951755