Fixing regression where icon view became focusable
Description: in ag/24721149 we stopped inflating IconView
from Task.xml and instead used a viewstub. When creating
files to fill the viewstub, we forgot to place the correct
attributes on Icon_view.xml. We did place them correctly
elsewhere. Anyways this is why we were seeing focusable
icons.

Flag: None

Test: Built locally, will follow up with automation

Bug: 307461880

Change-Id: I45f895c8545a80f80c79709180f2878476d10221