Update TaplOpenCloseAllApps click position to dismiss All Apps
After enabling responsive grid all apps container for tablet got wider. The previous calculation to click outside the All Apps container for this test no longer works.
This CL updates the click position to click after the All Apps scrollbar position. Also it changes the click to be right next to the container and not halfway to the edge. The calculation to get the position halfway to the edge was leading to a invalid click position where the click was happening inside the 24dp limit of the edge, and it was not dismissing the AllApps container.

Fix: 315076033
Fix: 315084703
Flag: ACONFIG com.android.launcher3.enable_responsive_workspace TEAMFOOD
Test: TaplOpenCloseAllApps
Test: TaplTestsTaskbar
Change-Id: I38bcdb4e0d4c27f35a16fa04411c65627bc88b11