Select task is broken on display large size
The update of the focus task overlay should be at last, cause otherwise the first time the "Select" button doesn't work - the text is not highlighted.

Also, we need to update pivot of the selected task only if there is a grid overview enabled. Otherwise the bug is still reproducible on Felix.

Bug: 305125147
Flag: NONE
Test: Manual (both on Tangor and Felix in Portrait/Landscape mode with small/medium/large Display size)
Change-Id: Id928cee0949b9124aeb27cc8210a53993e72ba19