Fixes crash when launching app pairs with intent+intent
This CL adds a new case for app pair launches. When we launch app pairs with intents, we receive initialTaskId and secondTaskId as -1, which was previously unhandled and caused a crash. With this change, we find the left/top app leash a different way, by looking for windowingMode=WINDOWING_MODE_MULTI_WINDOW and endAbsBounds.left/top == 0.

Fixes: 316050315
Test: Clear all running tasks from Overview, launch app pair, no longer crashes
Flag: ACONFIG com.android.wm.shell.enable_app_pairs DEVELOPMENT
Change-Id: Ic356d71ad267f079242213ebc59322bf1fb86b7c