Do not recreate taskbar due to unestimated bounds
The change around configuration and display cutout to support flexible
display setup is making the received display info is not calculated as
the hard-coded way in taskbar. It will cause the taskbar recreated when
the device reaches a given rotation for the first time.

The recreation is not necessary as it is only a hint of taskbar's
estimation doesn't match the result. Block the recreation in that case
to avoid user visible animation issue.

Bug: 302387383
Test: Rotate a device with movable cutout and no recreation happens
Flag: ACONFIG com.android.window.flags.allows_screen_size_decoupled_from_status_bar_and_cutout TRUNKFOOD
Flag: ACONFIG com.android.window.flags.movable_cutout_configuration DEVELOPMENT
Change-Id: I1aa6add57ec49a49cc7473bfaada6d9212c1fc4b