Use fade-in launcher for app pair launches from in-app taskbar
This CL temporarily patches a crash in this situation:

Inside of an app > Launch an unrelated app pair from Taskbar > Crash

The crash occurs because TaskbarActivityContext can't create a launcher animation (because Launcher isn't running). A different launch animation is needed for this case (similar to single apps), but for now, falling back to the fade-in launch at least allows the user to proceed without a crash.

Bug: 316485863
Test: Enter an app, launch an unrelated app pair, no crash
Flag: ACONFIG com.android.wm.shell.enable_app_pairs DEVELOPMENT
Change-Id: I72b2956b5b44616367e24263f2287824ee1e8212