Update live tile during dismiss anim even if disabled at start
Before we only checked if live tile drawing was enabled at the
start of the dismiss animation, and didn't add the frame callback
listener at all if not. Now we always add the frame callback
listener when dismissing the running task, and check if live tile
drawing is enabled within the listener in case there is a race.

Fixes: 312261719
Test: Swipe up to Overview from an app, immediately swipe up to
dismiss the app; ensure it always follows your finger
Flag: None

Change-Id: Ia12b02d5989c09d2104edd202f8c016f202f76b8