Perform TaskView animation when reopening the current task
When using gesture navigation to switch task, if the next started
task moves the previous task in live tile to top, the appeared
task will be the previous animating task rather than the task
started by recents.

Previously, none of condition handles the case, which causes a
splash screen task view to stay on screen and the transition keeps
in running state.

By running the side task launch animation, it can indicate that
there is an additional task switch visually, and the recents
transition can be finished when the animation is finished.

Bug: 317160303
Test: Activity A's onResume always starts activity B (declares
      different taskAffinity) with FLAG_ACTIVITY_NEW_TASK.
      Use gesture to swipe from B to A. B should be visible
      instead of staying at recents with a splash screen view.

Change-Id: Id1a441561ac9958f71efe2440e4fb467bedc6947