Making testAllAppsFromHome platinum
This test has a search for Youtube, which makes it possible for this test to replace android.platform.test.scenario.launcher.integration_tests.ScrollDownToSearchYoutubeFromAllApps#scrollDownToYouTube

Removing ScrollDownToSearchYoutubeFromAllApps#scrollDownToYouTube would move close to a state where the only Platinum tests outside of Launcher projects are ones that use 1P apps. This will make it easier to divide responsibilities between the IW and the Launcher teams.

Test: presubmit
Flag: N/A
Bug: 187761685
Change-Id: I889b1b23e5fe0a543eba24a922dc6908ac887305