Reduce the iconSize to fit inside the cellSize when display size largest
For some cases, even after reducing the iconSize and iconTextSize the content of the cell does not fit inside the cell height. This issue leads the icons or labels of the folders and workspace to crop. To prevent the icons of cropping when the display size or font size is largest (for acessibility) we allow the icons and labels shrink until it fits inside the defined cell height.

Bug: 314285621
Bug: 316087877
Flag: ACONFIG com.android.launcher3.enable_responsive_workspace TEAMFOOD
Test: NexusLauncherImageTests
Change-Id: Id8da3ee89658b9fc53a36d8605b9bd56ce168834