Simplify/fix testThreeButtonsTaskbarBoundsAfterConfigChangeDuringIme
Instead of dealing with changing dark theme which can lead to races
with stale Taskbar icons, instead just go home and launch the IME
activity. This is sufficient because the taskbar height changes
when going between home and an app, and the IME will stash the
Taskbar during this transition.

Fixes: 320490387
Test: testThreeButtonsTaskbarBoundsAfterConfigChangeDuringIme
Flag: None
Change-Id: Ib5b1481751af0bf1fccda085c78174f6612441b9