Skip bubble bar arrow animation if arrow detached
Arrow can become detached from the bubble bar background. This can
happen when bubble bar is pinned to the left and the right-most bubble
is selected. When the selected, right-most bubble is removed by drag and
dismiss, the view is removed and bubble bar background is immediately
updated. But we try to animate the arrow to the new selected position.
Before starting to animate arrow position, check if arrow position is
greater than bubble bar width. If it is, the arrow is detached and we
should not try to animate it to the new position. Just move it.

Flag: LEGACY persist.wm.debug.bubble_bar DEVELOPMENT
Bug: 273310265
Test: switch to RTL language, have 3 bubbles in bar, expand and select
  the right-most bubble, then dismiss it via drag, observe arrow jumps
  to new position and does not become detached from background
Test: switch to LTR language, try the same, 3 bubbles and dismiss the
  left-most bubble, observe arrow jumps the same and does not become
  detached
Test: select other bubbles, observe arrow position animates to the new
  position
Change-Id: I399db3c3e32b65b9125262bba0a0c32e45865bba