Move bubble bar stashed handle to the left for RTL
When system language is set to an RTL language, we move the bubble bar
to the left.
Move the stashed bubble bar handle to the left in this case.
Tweak the bubble bar expand animation so that the bubbles expand from
the left edge.

Flag: LEGACY persist.wm.debug.bubble_bar DEVELOPMENT
Bug: 273310265
Test: repeat following steps with LTR and RTL system languages
  - add couple of bubbles to bubble bar, open an app, observe stashed
    handle on the left
  - swipe up from system navbar to show taskbar and bubble bar
  - swipe up from bubble bar handle to show bubble bar and expand it
  - drag bubble bar to dismiss target and back out and release
Change-Id: Ia3b7892bba732a5278168d63db0594087f55fad4