Allow gesture tutorial to work for both orientations for square aspect ratios.
Devices with a square aspect ratio were previously seeing a prompt to rotate the screen for the gesture navigation tutorial. This change enables the tutorial for such devices but will still show the prompt to rotate the device if the aspect ratio would make the tutorial animations look wrong.

Flag: LEGACY ENABLE_NEW_GESTURE_NAV_TUTORIAL ENABLED
Fix: 321650001
Test: Tried gesture tutorial on various devices of different aspect
ratios and ensured that the prompt to rotate the device only shows up on
devices that do not have a square aspect ratio.

Change-Id: I7a6ca40e09a8202a9b54d898e0c68cc142694342